requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
aioredis==2.0.1
aiofiles==23.2.1
pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2